refactor(daxus): rename Accessor to BaseAccessor
