+++ packages/daxus/src/lib/model/index.ts
@@ -2,5 +2,5 @@ export * from './InfiniteAccessor.js';
 export * from './Model.js';
 export * from './NormalAccessor.js';
 export * from './types.js';
-export * from './Accessor.js';
+export * from './BaseAccessor.js';
 export * from './createDatabase.js';
